Update BlogNavHost : Seperated from MainActivity
